% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{Account}
\alias{Account}
\title{AdSense Host API Objects 
Generates performance reports, generates ad codes, and provides publisher management capabilities for AdSense Hosts.}
\usage{
Account(id = NULL, name = NULL, status = NULL)
}
\arguments{
\item{id}{Unique identifier of this account}

\item{name}{Name of this account}

\item{status}{Approval status of this account}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:33:46
filename: /Users/mark/dev/R/autoGoogleAPI/googleadsensehostv41.auto/R/adsensehost_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

