% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{AdClients}
\alias{AdClients}
\title{AdClients Object}
\usage{
AdClients(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The ad clients returned in this list response}

\item{nextPageToken}{Continuation token used to page through ad clients}
}
\value{
AdClients object
}
\description{
AdClients Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
