% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_functions.R
\docType{package}
\name{adsensehost_googleAuthR}
\alias{adsensehost_googleAuthR}
\alias{adsensehost_googleAuthR-package}
\title{AdSense Host API
Generates performance reports, generates ad codes, and provides publisher management capabilities for AdSense Hosts.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:24:20
filename: /Users/mark/dev/R/autoGoogleAPI/googleadsensehostv41.auto/R/adsensehost_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/adsensehost
}
}
