% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{AdStyle.colors}
\alias{AdStyle.colors}
\title{AdStyle.colors Object}
\usage{
AdStyle.colors(background = NULL, border = NULL, text = NULL,
  title = NULL, url = NULL)
}
\arguments{
\item{background}{The color of the ad background}

\item{border}{The color of the ad border}

\item{text}{The color of the ad text}

\item{title}{The color of the ad title}

\item{url}{The color of the ad url}
}
\value{
AdStyle.colors object
}
\description{
AdStyle.colors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The colors included in the style. These are represented as six hexadecimal characters, similar to HTML color codes, but without the leading hash.
}
\seealso{
Other AdStyle functions: \code{\link{AdStyle.font}},
  \code{\link{AdStyle}}
}
