% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{AdStyle.font}
\alias{AdStyle.font}
\title{AdStyle.font Object}
\usage{
AdStyle.font(family = NULL, size = NULL)
}
\arguments{
\item{family}{The family of the font}

\item{size}{The size of the font}
}
\value{
AdStyle.font object
}
\description{
AdStyle.font Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The font which is included in the style.
}
\seealso{
Other AdStyle functions: \code{\link{AdStyle.colors}},
  \code{\link{AdStyle}}
}

