% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{AdUnit.contentAdsSettings.backupOption}
\alias{AdUnit.contentAdsSettings.backupOption}
\title{AdUnit.contentAdsSettings.backupOption Object}
\usage{
AdUnit.contentAdsSettings.backupOption(color = NULL, type = NULL,
  url = NULL)
}
\arguments{
\item{color}{Color to use when type is set to COLOR}

\item{type}{Type of the backup option}

\item{url}{URL to use when type is set to URL}
}
\value{
AdUnit.contentAdsSettings.backupOption object
}
\description{
AdUnit.contentAdsSettings.backupOption Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The backup option to be used in instances where no ad is available.
}
\seealso{
Other AdUnit functions: \code{\link{AdUnit.contentAdsSettings}},
  \code{\link{AdUnit.mobileContentAdsSettings}},
  \code{\link{AdUnit}}
}

