% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{AssociationSession}
\alias{AssociationSession}
\title{AssociationSession Object}
\usage{
AssociationSession(accountId = NULL, id = NULL, productCodes = NULL,
  redirectUrl = NULL, status = NULL, userLocale = NULL,
  websiteLocale = NULL, websiteUrl = NULL)
}
\arguments{
\item{accountId}{Hosted account id of the associated publisher after association}

\item{id}{Unique identifier of this association session}

\item{productCodes}{The products to associate with the user}

\item{redirectUrl}{Redirect URL of this association session}

\item{status}{Status of the completed association, available once the association callback token has been verified}

\item{userLocale}{The preferred locale of the user themselves when going through the AdSense association flow}

\item{websiteLocale}{The locale of the user's hosted website}

\item{websiteUrl}{The URL of the user's hosted website}
}
\value{
AssociationSession object
}
\description{
AssociationSession Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

