% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{CustomChannel}
\alias{CustomChannel}
\title{CustomChannel Object}
\usage{
CustomChannel(code = NULL, id = NULL, name = NULL)
}
\arguments{
\item{code}{Code of this custom channel, not necessarily unique across ad clients}

\item{id}{Unique identifier of this custom channel}

\item{name}{Name of this custom channel}
}
\value{
CustomChannel object
}
\description{
CustomChannel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CustomChannel functions: \code{\link{customchannels.insert}},
  \code{\link{customchannels.patch}},
  \code{\link{customchannels.update}}
}
