% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{CustomChannels}
\alias{CustomChannels}
\title{CustomChannels Object}
\usage{
CustomChannels(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The custom channels returned in this list response}

\item{nextPageToken}{Continuation token used to page through custom channels}
}
\value{
CustomChannels object
}
\description{
CustomChannels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
