% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_objects.R
\name{Report}
\alias{Report}
\title{Report Object}
\usage{
Report(Report.headers = NULL, averages = NULL, headers = NULL,
  rows = NULL, totalMatchedRows = NULL, totals = NULL, warnings = NULL)
}
\arguments{
\item{Report.headers}{The \link{Report.headers} object or list of objects}

\item{averages}{The averages of the report}

\item{headers}{The header information of the columns requested in the report}

\item{rows}{The output rows of the report}

\item{totalMatchedRows}{The total number of rows matched by the report request}

\item{totals}{The totals of the report}

\item{warnings}{Any warnings associated with generation of the report}
}
\value{
Report object
}
\description{
Report Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Report functions: \code{\link{Report.headers}}
}

