% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_functions.R
\name{reports.generate}
\alias{reports.generate}
\title{Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify 'alt=csv' as a query parameter.}
\usage{
reports.generate(startDate, endDate, dimension = NULL, filter = NULL,
  locale = NULL, maxResults = NULL, metric = NULL, sort = NULL,
  startIndex = NULL)
}
\arguments{
\item{startDate}{Start of the date range to report on in 'YYYY-MM-DD' format, inclusive}

\item{endDate}{End of the date range to report on in 'YYYY-MM-DD' format, inclusive}

\item{dimension}{Dimensions to base the report on}

\item{filter}{Filters to be run on the report}

\item{locale}{Optional locale to use for translating report output to a local language}

\item{maxResults}{The maximum number of rows of report data to return}

\item{metric}{Numeric columns to include in the report}

\item{sort}{The name of a dimension or metric to sort the resulting report on, optionally prefixed with '+' to sort ascending or '-' to sort descending}

\item{startIndex}{Index of the first row of report data to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adsensehost
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adsensehost)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/adsense/host/}{Google Documentation}
}

