% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsensehost_functions.R
\name{urlchannels.insert}
\alias{urlchannels.insert}
\title{Add a new URL channel to the host AdSense account.}
\usage{
urlchannels.insert(UrlChannel, adClientId)
}
\arguments{
\item{UrlChannel}{The \link{UrlChannel} object to pass to this method}

\item{adClientId}{Ad client to which the new URL channel will be added}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adsensehost
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adsensehost)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/adsense/host/}{Google Documentation}

Other UrlChannel functions: \code{\link{UrlChannel}}
}
